// web/src/modules/defang/generated/fabric_pb.ts
export enum SubscriptionTier {
  /**
   * @generated from enum value: SUBSCRIPTION_TIER_UNSPECIFIED = 0;
   */
  SUBSCRIPTION_TIER_UNSPECIFIED = 0,

  /**
   * @generated from enum value: HOBBY = 1;
   */
  HOBBY = 1,

  /**
   * @generated from enum value: PERSONAL = 2;
   */
  PERSONAL = 2,

  /**
   * @generated from enum value: PRO = 3;
   */
  PRO = 3,

  /**
   * @generated from enum value: TEAM = 4;
   */
  TEAM = 4,
}